Outsourced Configuration setup and setting defaults to be used by tests
